Move Seq2C prep_bed to clean_inputs to avoid concurrency issue.
