+++ bcbio/cwl/defs.py
@@ -138,18 +138,21 @@ def variant():
                 ["reference", "fasta", "base"]],
                [cwlout(["config", "algorithm", "variant_regions"], ["File", "null"]),
                 cwlout(["config", "algorithm", "variant_regions_merged"], ["File", "null"]),
-                cwlout(["config", "algorithm", "variant_regions_orig"], ["File", "null"])]),
+                cwlout(["config", "algorithm", "variant_regions_orig"], ["File", "null"]),
+                cwlout(["config", "algorithm", "seq2c_bed_ready"], ["File", "null"])]),
              s("postprocess_alignment", "multi-parallel",
                [["align_bam"], ["config", "algorithm", "variant_regions"],
                 ["config", "algorithm", "coverage_interval"],
                 ["config", "algorithm", "variant_regions_merged"],
                 ["config", "algorithm", "variant_regions_orig"],
+                ["config", "algorithm", "seq2c_bed_ready"],
                 ["config", "algorithm", "recalibrate"],
                 ["reference", "fasta", "base"]],
                [cwlout(["config", "algorithm", "coverage_interval"], "string"),
                 cwlout(["config", "algorithm", "variant_regions"], "File"),
                 cwlout(["config", "algorithm", "variant_regions_merged"], "File"),
                 cwlout(["config", "algorithm", "variant_regions_orig"], "File"),
+                cwlout(["config", "algorithm", "seq2c_bed_ready"], "File"),
                 cwlout(["regions", "callable"], "File"),
                 cwlout(["regions", "sample_callable"], "File"),
                 cwlout(["regions", "nblock"], "File"),
