+++ bcbio/structural/seq2c.py
@@ -19,7 +19,7 @@ from bcbio.provenance import do
 from bcbio.structural import annotate, regions
 from bcbio.log import logger
 from bcbio.variation.coverage import regions_coverage
-from bcbio.variation.bedutils import clean_file
+from bcbio.variation import bedutils
 
 
 def precall(items):
@@ -35,7 +35,7 @@ def precall(items):
     # assert dd.get_coverage_interval(data) != "genome", "Seq2C only for amplicon and exome sequencing"
 
     work_dir = _sv_workdir(data)
-    bed_file = _prep_bed(data, work_dir)
+    bed_file = data["config"]["algorithm"]["seq2c_bed_ready"]
     bam_file = dd.get_align_bam(data)
     sample_name = dd.get_sample_name(data)
 
@@ -67,14 +67,14 @@ def run(items):
 
     return items
 
-def _prep_bed(data, work_dir):
+def prep_seq2c_bed(data):
     """Selecting the bed file, cleaning, and properly annotating for Seq2C
     """
     bed_file = regions.get_sv_bed(data)
     if bed_file:
-        bed_file = clean_file(bed_file, data, prefix="svregions-")
+        bed_file = bedutils.clean_file(bed_file, data, prefix="svregions-")
     else:
-        bed_file = clean_file(dd.get_variant_regions(data), data)
+        bed_file = bedutils.clean_file(dd.get_variant_regions(data), data)
 
     col_num = bt.BedTool(bed_file).field_count()
     if col_num < 4:
